+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Dev
+
+- Detect if a linker is installed and don't display warning if it is (https://github.com/Kobzol/cargo-wizard/issues/5).
+
 # 0.2.1 (10. 3. 2024)
 
 - Add Unix-specific options.